News_Ingestor/requirements.txt
# news_ingestor/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
gunicorn==22.0.0
httpx==0.27.0 # For asynchronous HTTP requests
google-cloud-pubsub==2.21.0
python-dotenv==1.0.1 # For local development
